More header fixes.
I just realized @nyanpasu asked me to do the rest.

But now I'm wondering if it really was worth it.
If we take a look, the build now uses way less ressource, since the compilation of each object file requires
a minimal amount of headers, this makes objects and their heider not interdepandant and thus using about half the RAM it used to.
Travis will probably won't have to kill g++ from now.

On the other hand, even if less memory is used at a given time, the memory usage is still the same but scattered over time, resulting in theorically slower builds.
For example, when compiling GtkMainWindow, and GtkSettingDialog, the gtkmm/button.h header is going to be parsed twice. While if it had been included in a common header,
the include guards would have prevented the redefinition and the linker would have done its job, not sure if it actually works like this, it's 6 am and brain is damaged. /bullshit

Anyway I doubt we'll have travis crashes from now.