More RSS Dialogs.
The add rss group in the toolbar finally properly popups the dialog.
You double click it in the sidebar, enter a name, then you'll be prompted for settings.
I wrote a few comments on how it will work but here's a sum up.

You have an RSS group, made of multiple feeds.
All these feeds get their items mixed in a big list, and on each item a regex is used to
compare the matching string or number to a fixed value. When the comparison is true, the item is marked ready for download.
When the user clicks on ok, all torrent in the items list starts downloading, and the user choose if for every new item,  if he wants to be notified
 or start the download immediatly.

This is how RSS is planned to be implemented.
I still need to make a few changes in wip/rss in core since I haven't planned for groups to have multiple feeds in there.