Implement the RSS dialog
Also note that the Feed Name field when you add a new feed is unused,
it's overwritten by the feed title specified in the RSS XML, or the
 url, I'll remove it later.

Currently auto adding item is disabled, mainly because I didn't
want to accidentally allocate hundreds of torrents at once.

What's left to do is:

*Add a global setting asking if the user wants to be notified
 of a new available item
*Add a per-group basis settings to auto download new items.
*Show the number of downloading items next to the group name in
 the sidebar (like "group_name (5)")
*A single click on the group should show in the torrent treeview only
  the downloading/seeding items of the group
*Automatically update the various treeviews in the RSS dialog
*Color in green in the rssTreeview the items that passed the filters,
 in red the others

*Labels
*A LOT OF TESTING

After this, ui/rewrite will be merged in master. We'll do the
add torrent dialog (so you can pick and choose files)

And finally I'll take a break from -gtk and focus on windows.